🎨 Split user service
